+++ CHANGELOG.md
@@ -3,11 +3,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [UNRELEASED]
+## [0.12.0] - 2022-02-19
 ### Added
 - support for colored header
 - custom scrollbars (opt-in)
 - `Drawer` component
+- `Stats` component
+- dots icon
 ### Changed
 - added a prop for controlling closing `Pop` on `Escape` key
 - added className for sticky header content component